fixed spacing and style issues
