+++ app/javascript/packs/components/QuestionnaireControls.jsx
@@ -5,7 +5,7 @@
 
 import React from 'react';
 
-import {Link} from 'react-router-dom';
+import { Link } from 'react-router-dom';
 
 class QuestionnaireControls extends React.Component {
   render() {
