+++ app/javascript/packs/containers/NavigationBar/CategoryContainer.jsx
@@ -1,4 +1,4 @@
-import {Link} from 'react-router-dom'
+import { Link } from 'react-router-dom'
 import React from 'react'
