+++ app/javascript/packs/containers/PortfolioListContainer.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import {loadInitialState} from '../actions/initialState';
-import {bindActionCreators} from 'redux';
-import {connect} from 'react-redux';
-import {Link} from 'react-router-dom';
+import { loadInitialState } from '../actions/initialState';
+import { bindActionCreators } from 'redux';
+import { connect } from 'react-redux';
+import { Link } from 'react-router-dom';
 
 class PortfolioListContainer extends React.Component {
   componentDidMount() {
